+++ src/Pages/Shared/ConfigMapSecret/ConfigMapSecretProtected.tsx
@@ -16,9 +16,7 @@ import {
 } from '@devtron-labs/devtron-fe-common-lib'
 
 import { hasApproverAccess, importComponentFromFELibrary } from '@Components/common'
-import CompareConfigView from '@Pages/Applications/DevtronApps/Details/AppConfigurations/MainContent/CompareConfigView'
-import NoPublishedVersionEmptyState from '@Pages/Applications/DevtronApps/Details/AppConfigurations/MainContent/NoPublishedVersionEmptyState'
-import { CompareConfigViewProps } from '@Pages/Applications/DevtronApps/Details/AppConfigurations/MainContent/types'
+import { CompareConfigView, CompareConfigViewProps, NoPublishedVersionEmptyState } from '@Pages/Applications'
 
 import { CM_SECRET_STATE, CMSecretComponentType, CMSecretConfigData, ConfigMapSecretProtectedProps } from './types'
 import { getConfigMapSecretPayload, getConfigMapSecretReadOnlyValues } from './utils'
